multi-value query parameters

+++ services/mgmt/batch/src/package_2017_09/operations.rs
@@ -1078,7 +1078,7 @@ pub mod application {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         account_name: &str,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ListApplicationsResult, list::Error> {
         let client = &operation_config.client;
@@ -1334,7 +1334,7 @@ pub mod certificate {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         account_name: &str,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         select: Option<&str>,
         filter: Option<&str>,
         subscription_id: &str,
@@ -1743,7 +1743,7 @@ pub mod pool {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         account_name: &str,
-        maxresults: Option<i64>,
+        maxresults: Option<i32>,
         select: Option<&str>,
         filter: Option<&str>,
         subscription_id: &str,
