+++ services/mgmt/devtestlabs/src/package_2018_09/operations.rs
@@ -55,7 +55,7 @@ pub mod labs {
         subscription_id: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<LabList, list_by_subscription::Error> {
         let client = &operation_config.client;
@@ -131,7 +131,7 @@ pub mod labs {
         resource_group_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<LabList, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -926,7 +926,7 @@ pub mod global_schedules {
         subscription_id: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ScheduleList, list_by_subscription::Error> {
         let client = &operation_config.client;
@@ -1002,7 +1002,7 @@ pub mod global_schedules {
         resource_group_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ScheduleList, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -1476,7 +1476,7 @@ pub mod artifact_sources {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ArtifactSourceList, list::Error> {
         let client = &operation_config.client;
@@ -1827,7 +1827,7 @@ pub mod arm_templates {
         artifact_source_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ArmTemplateList, list::Error> {
         let client = &operation_config.client;
@@ -1977,7 +1977,7 @@ pub mod artifacts {
         artifact_source_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ArtifactList, list::Error> {
         let client = &operation_config.client;
@@ -2333,7 +2333,7 @@ pub mod custom_images {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<CustomImageList, list::Error> {
         let client = &operation_config.client;
@@ -2685,7 +2685,7 @@ pub mod formulas {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<FormulaList, list::Error> {
         let client = &operation_config.client;
@@ -3035,7 +3035,7 @@ pub mod gallery_images {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<GalleryImageList, list::Error> {
         let client = &operation_config.client;
@@ -3117,7 +3117,7 @@ pub mod notification_channels {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<NotificationChannelList, list::Error> {
         let client = &operation_config.client;
@@ -3598,7 +3598,7 @@ pub mod policies {
         policy_set_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<PolicyList, list::Error> {
         let client = &operation_config.client;
@@ -3952,7 +3952,7 @@ pub mod schedules {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ScheduleList, list::Error> {
         let client = &operation_config.client;
@@ -4633,7 +4633,7 @@ pub mod users {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<UserList, list::Error> {
         let client = &operation_config.client;
@@ -4986,7 +4986,7 @@ pub mod disks {
         user_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<DiskList, list::Error> {
         let client = &operation_config.client;
@@ -5477,7 +5477,7 @@ pub mod environments {
         user_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<DtlEnvironmentList, list::Error> {
         let client = &operation_config.client;
@@ -5834,7 +5834,7 @@ pub mod secrets {
         user_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<SecretList, list::Error> {
         let client = &operation_config.client;
@@ -6189,7 +6189,7 @@ pub mod service_fabrics {
         user_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ServiceFabricList, list::Error> {
         let client = &operation_config.client;
@@ -6738,7 +6738,7 @@ pub mod service_fabric_schedules {
         service_fabric_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ScheduleList, list::Error> {
         let client = &operation_config.client;
@@ -7162,7 +7162,7 @@ pub mod virtual_machines {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<LabVirtualMachineList, list::Error> {
         let client = &operation_config.client;
@@ -8352,7 +8352,7 @@ pub mod virtual_machine_schedules {
         virtual_machine_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ScheduleList, list::Error> {
         let client = &operation_config.client;
@@ -8771,7 +8771,7 @@ pub mod virtual_networks {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<VirtualNetworkList, list::Error> {
         let client = &operation_config.client;
