+++ services/mgmt/intune/src/package_2015_01_privatepreview/operations.rs
@@ -93,7 +93,7 @@ pub async fn get_apps(
     operation_config: &crate::OperationConfig,
     host_name: &str,
     filter: Option<&str>,
-    top: Option<i64>,
+    top: Option<i32>,
     select: Option<&str>,
 ) -> std::result::Result<ApplicationCollection, get_apps::Error> {
     let client = &operation_config.client;
@@ -153,7 +153,7 @@ pub async fn get_mam_user_devices(
     host_name: &str,
     user_name: &str,
     filter: Option<&str>,
-    top: Option<i64>,
+    top: Option<i32>,
     select: Option<&str>,
 ) -> std::result::Result<DeviceCollection, get_mam_user_devices::Error> {
     let client = &operation_config.client;
@@ -318,7 +318,7 @@ pub async fn get_operation_results(
     operation_config: &crate::OperationConfig,
     host_name: &str,
     filter: Option<&str>,
-    top: Option<i64>,
+    top: Option<i32>,
     select: Option<&str>,
 ) -> std::result::Result<OperationResultCollection, get_operation_results::Error> {
     let client = &operation_config.client;
@@ -425,7 +425,7 @@ pub async fn get_mam_flagged_users(
     operation_config: &crate::OperationConfig,
     host_name: &str,
     filter: Option<&str>,
-    top: Option<i64>,
+    top: Option<i32>,
     select: Option<&str>,
 ) -> std::result::Result<FlaggedUserCollection, get_mam_flagged_users::Error> {
     let client = &operation_config.client;
@@ -541,7 +541,7 @@ pub async fn get_mam_user_flagged_enrolled_apps(
     host_name: &str,
     user_name: &str,
     filter: Option<&str>,
-    top: Option<i64>,
+    top: Option<i32>,
     select: Option<&str>,
 ) -> std::result::Result<FlaggedEnrolledAppCollection, get_mam_user_flagged_enrolled_apps::Error> {
     let client = &operation_config.client;
@@ -608,7 +608,7 @@ pub mod ios {
         operation_config: &crate::OperationConfig,
         host_name: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         select: Option<&str>,
     ) -> std::result::Result<IosmamPolicyCollection, get_mam_policies::Error> {
         let client = &operation_config.client;
@@ -875,7 +875,7 @@ pub mod ios {
         host_name: &str,
         policy_name: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         select: Option<&str>,
     ) -> std::result::Result<ApplicationCollection, get_app_for_mam_policy::Error> {
         let client = &operation_config.client;
@@ -1200,7 +1200,7 @@ pub mod android {
         operation_config: &crate::OperationConfig,
         host_name: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         select: Option<&str>,
     ) -> std::result::Result<AndroidMamPolicyCollection, get_mam_policies::Error> {
         let client = &operation_config.client;
@@ -1468,7 +1468,7 @@ pub mod android {
         host_name: &str,
         policy_name: &str,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         select: Option<&str>,
     ) -> std::result::Result<ApplicationCollection, get_app_for_mam_policy::Error> {
         let client = &operation_config.client;
