+++ services/mgmt/keyvault/src/package_preview_2020_04/operations.rs
@@ -273,7 +273,7 @@ pub mod vaults {
     pub async fn list_by_resource_group(
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<VaultListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -320,7 +320,7 @@ pub mod vaults {
     }
     pub async fn list_by_subscription(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<VaultListResult, list_by_subscription::Error> {
         let client = &operation_config.client;
@@ -499,7 +499,7 @@ pub mod vaults {
     pub async fn list(
         operation_config: &crate::OperationConfig,
         filter: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list::Error> {
         let client = &operation_config.client;
@@ -1175,7 +1175,7 @@ pub mod managed_hsms {
     pub async fn list_by_resource_group(
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ManagedHsmListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -1240,7 +1240,7 @@ pub mod managed_hsms {
     }
     pub async fn list_by_subscription(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ManagedHsmListResult, list_by_subscription::Error> {
         let client = &operation_config.client;
