+++ services/mgmt/keyvault/src/package_preview_2020_04_full/operations.rs
@@ -273,7 +273,7 @@ pub mod vaults {
     pub async fn list_by_resource_group(
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<VaultListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -320,7 +320,7 @@ pub mod vaults {
     }
     pub async fn list_by_subscription(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<VaultListResult, list_by_subscription::Error> {
         let client = &operation_config.client;
@@ -499,7 +499,7 @@ pub mod vaults {
     pub async fn list(
         operation_config: &crate::OperationConfig,
         filter: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ResourceListResult, list::Error> {
         let client = &operation_config.client;
@@ -1063,7 +1063,7 @@ pub mod secrets {
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
         vault_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<SecretListResult, list::Error> {
         let client = &operation_config.client;
@@ -1387,7 +1387,7 @@ pub mod managed_hsms {
     pub async fn list_by_resource_group(
         operation_config: &crate::OperationConfig,
         resource_group_name: &str,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ManagedHsmListResult, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -1452,7 +1452,7 @@ pub mod managed_hsms {
     }
     pub async fn list_by_subscription(
         operation_config: &crate::OperationConfig,
-        top: Option<i64>,
+        top: Option<i32>,
         subscription_id: &str,
     ) -> std::result::Result<ManagedHsmListResult, list_by_subscription::Error> {
         let client = &operation_config.client;
