+++ services/mgmt/labservices/src/package_2018_10/operations.rs
@@ -679,7 +679,7 @@ pub mod lab_accounts {
         subscription_id: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ResponseWithContinuationLabAccount, list_by_subscription::Error> {
         let client = &operation_config.client;
@@ -756,7 +756,7 @@ pub mod lab_accounts {
         resource_group_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ResponseWithContinuationLabAccount, list_by_resource_group::Error> {
         let client = &operation_config.client;
@@ -1290,7 +1290,7 @@ pub mod gallery_images {
         lab_account_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ResponseWithContinuationGalleryImage, list::Error> {
         let client = &operation_config.client;
@@ -1641,7 +1641,7 @@ pub mod labs {
         lab_account_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ResponseWithContinuationLab, list::Error> {
         let client = &operation_config.client;
@@ -2110,7 +2110,7 @@ pub mod environment_settings {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ResponseWithContinuationEnvironmentSetting, list::Error> {
         let client = &operation_config.client;
@@ -2717,7 +2717,7 @@ pub mod environments {
         environment_setting_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ResponseWithContinuationEnvironment, list::Error> {
         let client = &operation_config.client;
@@ -3310,7 +3310,7 @@ pub mod users {
         lab_name: &str,
         expand: Option<&str>,
         filter: Option<&str>,
-        top: Option<i64>,
+        top: Option<i32>,
         orderby: Option<&str>,
     ) -> std::result::Result<ResponseWithContinuationUser, list::Error> {
         let client = &operation_config.client;
